requirements.txt
Adafruit-Blinka==8.47.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-framebuf==1.6.5
adafruit-circuitpython-pn532==2.4.1
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-ssd1306==2.12.17
adafruit-circuitpython-typing==1.11.0
Adafruit-PlatformDetect==3.74.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorzero==2.0
cryptography==43.0.1
firebase_admin==6.5.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gpiozero==2.0.1
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyftdi==0.55.4
pygame==2.6.1
PyJWT==2.9.0
pyparsing==3.1.4
pyserial==3.5
pyusb==1.2.1
requests==2.32.3
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
RPLCD==1.3.1
rsa==4.9
sysv-ipc==1.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
